Move the verification of QIR profile out to its own file. Update the pass to use the newer tablegen features.
